+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Rule::name` method
 
+### Removed
+
+- `Value::new` constructor
+
+
 
 ## [0.4.3]